back-end/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.2
werkzeug==2.2.2
keras
joblib
numpy
tensorflow
scikit-learn